# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Club Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /enrollments:
    get:
      operationId: getEnrollments
      tags: [enrollments]
      summary: Get all trimester enrollments
      parameters:
        - name: leadKey
          in: query
          schema:
            type: string
        - name: userUsername
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Create a trimester enrollment (Step 4)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    patch:
      operationId: updateEnrollment
      tags: [enrollments]
      summary: Update an enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/reroll-status:
    get:
      operationId: getRerollStatus
      tags: [enrollments]
      summary: List all users with their enrollment status for a given trimester/year
      parameters:
        - name: trimester
          in: query
          required: true
          schema:
            type: string
        - name: academicYear
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reroll status list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RerollCandidate"
  /enrollments/reroll:
    post:
      operationId: rerollEnrollments
      tags: [enrollments]
      summary: Bulk re-enroll selected users for a trimester/year
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RerollInput"
      responses:
        "200":
          description: Reroll result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RerollResult"
  /enrollments/{id}/send-receipt:
    post:
      operationId: sendEnrollmentReceipt
      tags: [enrollments]
      summary: Generate PDF receipt and send via WhatsApp
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Receipt sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReceiptResult"

  # Global Payments Dashboard
...
components:
  schemas:
...
    Enrollment:
      type: object
      required: [id, leadKey, leadSource]
      properties:
        id:
          type: string
        leadKey:
          type: string
        leadSource:
          type: string
        userUsername:
          type: ["string", "null"]
        program:
          type: string
        categorie:
          type: string
        paymentMethod:
          type: string
        groupName:
          type: string
        parcours:
          type: string
        horaire:
          type: string
        paymentStatus:
          type: string
        facilitator:
          type: string
        amountExpected:
          type: number
        amountReceived:
          type: number
        amountRemaining:
          type: number
        amountPerPeriod:
          type: ["number", "null"]
        billingStartDate:
          type: ["string", "null"]
        programStartDate:
          type: ["string", "null"]
        programEndDate:
          type: ["string", "null"]
        depositDueDate:
          type: ["string", "null"]
        suspensionStatus:
          type: string
        invoiceNumber:
          type: string
        trimester:
          type: string
        academicYear:
          type: string
        receiptSentAt:
          type: ["string", "null"]
        createdAt:
          type: string
    EnrollmentInput:
      type: object
      required: [leadKey, leadSource]
      properties:
        leadKey:
          type: string
        leadSource:
          type: string
        userUsername:
          type: string
        program:
          type: string
        categorie:
          type: string
        paymentMethod:
          type: string
        groupName:
          type: string
        parcours:
          type: string
        horaire:
          type: string
        facilitator:
          type: string
        amountExpected:
          type: number
        amountReceived:
          type: number
        amountPerPeriod:
          type: number
        billingStartDate:
          type: string
        programStartDate:
          type: string
        programEndDate:
          type: string
        depositDueDate:
          type: string
        suspensionStatus:
          type: string
        invoiceNumber:
          type: string
        trimester:
          type: string
        academicYear:
          type: string
    EnrollmentUpdate:
      type: object
      properties:
        userUsername:
          type: string
        program:
          type: string
        paymentMethod:
          type: string
        groupName:
          type: string
        parcours:
          type: string
        horaire:
          type: string
        facilitator:
          type: string
        amountExpected:
          type: number
        amountReceived:
          type: number
        amountPerPeriod:
          type: ["number", "null"]
        billingStartDate:
          type: string
        programStartDate:
          type: string
        programEndDate:
          type: string
        depositDueDate:
          type: string
        suspensionStatus:
          type: string
        invoiceNumber:
          type: string
        trimester:
          type: string
        academicYear:
          type: string
    ReceiptResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        error:
          type: string
...
    RerollCandidate:
      type: object
      required: [username, nom, prenom, enrolled]
      properties:
        username:
          type: string
        nom:
          type: string
        prenom:
          type: string
        categorie:
          type: string
        groupName:
          type: string
        parcours:
          type: string
        horaire:
          type: string
        enrolled:
          type: boolean
        enrollmentId:
          type: string
          nullable: true
        program:
          type: string
        paymentMethod:
          type: string
        amountPerPeriod:
          type: ["number", "null"]
    RerollEntryInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
        program:
          type: string
        paymentMethod:
          type: string
        amountPerPeriod:
          type: number
        groupName:
          type: string
        parcours:
          type: string
        horaire:
          type: string
    RerollInput:
      type: object
      required: [trimester, academicYear, entries]
      properties:
        trimester:
          type: string
        academicYear:
          type: string
        entries:
          type: array
          items:
            $ref: "#/components/schemas/RerollEntryInput"
    RerollResult:
      type: object
      required: [success, enrolled]
      properties:
        success:
          type: boolean
        enrolled:
          type: integer
        enrollments:
          type: array
          items:
            $ref: "#/components/schemas/Enrollment"
...